[move] #6 TheaterSelectionModalBottomSheet 따로 파일로 빼냄
